requirement.txt
# 核心库
fastapi==0.104.1          # Web框架
uvicorn[standard]==0.24.0 # ASGI服务器
pydantic==2.5.0           # 数据验证

# AI相关
langchain==0.1.0          # 大模型应用框架
chromadb==0.4.22          # 向量数据库
sentence-transformers==2.2.2  # 本地Embedding模型
openai==1.6.1             # OpenAI API（如果你用GPT）

# 文档处理
pypdf==3.17.0             # PDF解析
python-docx==1.1.0        # Word文档解析
markdown==3.5.1           # Markdown解析

# 工具
pyyaml==6.0.1             # YAML配置解析
python-dotenv==1.0.0      # 环境变量管理
tiktoken==0.5.1           # Token计数

# 前端（简单版）
streamlit==1.29.0         # 快速构建UI
